requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.4
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.1
langchain==0.1.13
langchain-cli==0.0.21
langchain-community==0.0.32
langchain-core==0.1.41
langchain-experimental==0.0.55
langchain-openai==0.1.0
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langserve==0.0.51
langsmith==0.1.31
logging==0.4.9.6
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
packaging==23.2
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pypdf==4.1.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
rapidfuzz==3.7.0
redis==5.0.3
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sse-starlette==1.8.2
starlette==0.36.3
tavily-python==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
trove-classifiers==2024.3.25
typer==0.9.4
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.23.2
virtualenv==20.25.1
xattr==1.1.0
yarl==1.9.4
zipp==3.18.1
